+++ lib/OrdinaryDiffEqDifferentiation/src/OrdinaryDiffEqDifferentiation.jl
@@ -10,9 +10,19 @@ import ForwardDiff, FiniteDiff
 import ForwardDiff.Dual
 import LinearSolve
 import LinearSolve: OperatorAssumptions
+import FunctionWrappersWrappers
 using DiffEqBase
+
 import LinearAlgebra
+import LinearAlgebra: Diagonal, I, UniformScaling, diagind
+import SparseArrays: SparseMatrixCSC
+
 import InteractiveUtils
+import ArrayInterface
+
+import StaticArrayInterface
+import StaticArrays: SArray, MVector, SVector, @SVector, StaticArray, MMatrix, SA,
+                     StaticMatrix
 
 using DiffEqBase: TimeGradientWrapper,
                   UJacobianWrapper, TimeDerivativeWrapper,
@@ -21,7 +31,8 @@ using SciMLBase: AbstractSciMLOperator
 using OrdinaryDiffEq: OrdinaryDiffEqAlgorithm, OrdinaryDiffEqAdaptiveImplicitAlgorithm, DAEAlgorithm,
 OrdinaryDiffEqImplicitAlgorithm, CompositeAlgorithm, OrdinaryDiffEqExponentialAlgorithm,
 OrdinaryDiffEqAdaptiveExponentialAlgorithm, @unpack, AbstractNLSolver, nlsolve_f, issplit,
-concrete_jac
+concrete_jac, unwrap_alg, OrdinaryDiffEqCache, _vec, standardtag, isnewton, _unwrap_val
+import OrdinaryDiffEq: get_chunksize
 
 import OrdinaryDiffEq: alg_autodiff
