+++ lib/OrdinaryDiffEqNonlinearSolve/src/utils.jl
@@ -15,7 +15,6 @@ isnewton(::AbstractNLSolverCache) = false
 isnewton(::Union{NLNewtonCache, NLNewtonConstantCache}) = true
 
 is_always_new(nlsolver::AbstractNLSolver) = is_always_new(nlsolver.alg)
-is_always_new(alg) = isdefined(alg, :always_new) ? alg.always_new : false
 check_div(nlsolver::AbstractNLSolver) = check_div(nlsolver.alg)
 check_div(alg) = isdefined(alg, :check_div) ? alg.check_div : true
